Fix "numpy" dependency in "setup.py"
# Issue

If "numpy" had not been installed, "pyoctree" was failing to build and
install when running "setup.py" script (e.g., with "pip"). That was
happening because of the "import numpy" statement. If there was no
"numpy" installed in advance, Python was failing at that point.

Also, while "install_requires" parameter was adjusted recently to define
"numpy" as an additional dependency, it needed to go to "setup_requires"
list as well.

# Fix

* Added "build_ext_dependencies" class to import "numpy" and add all the
  required dependencies only when it's actually needed
* Added "setup_requires" parameter to specify "numpy" as a dependency
  for building the project